+++ resources/views/marketing/legal/document.blade.php
@@ -1,0 +1,9 @@
+@extends('layouts.marketing')
+
+@section('title', $title . ' – GastroBook')
+
+@section('content')
+<section class="legal mx-auto max-w-3xl px-4 py-16">
+    {!! $html !!}
+</section>
+@endsection